chore(models): refactor models
